Move implementation out of the test file
