+++ experimental/src/superoptimizer.py
@@ -223,7 +223,11 @@ class CpuState(StateBase):
         if not isinstance(other, type(self)):
             return False
 
-        return (self.pipeline, self.computed_values, self.goal_exprs,) == (
+        return (
+            self.pipeline,
+            self.computed_values,
+            self.goal_exprs,
+        ) == (
             other.pipeline,
             other.computed_values,
             other.goal_exprs,
